backend/requirements.txt
django>=4.2,<5.0
psycopg2-binary  # Para o PostgreSQL
paho-mqtt        # Para o IoT (MQTT)
djangorestframework  # Para o Front em React
django-cors-headers  # Para o React conseguir falar com o Django
